// src/app/model/password-cracking-data.model.ts
export interface PasswordCrackingData {
    characters: number;
    numbersOnly: {
       value: string;
       unit: string;
       color: string;
    };
    lowercaseOnly: {
       value: string;
       unit: string;
       color: string;
    };
    mixedCase: {
       value: string;
       unit: string;
       color: string;
    };
    numbersAndLetters: {
       value: string;
       unit: string;
       color: string;
    };
    allCharacters: {
       value: string;
       unit: string;
       color: string;
    };
 }
 
 